Create stop_watch.py
my first to use the frame in codeskulptor 
the math in this one is simple, of course, just comparisons, you can see I listed every global with the word "global" before it in the same line, instead of just listing the variables. but it was one of the first codes I wrote.