Create blackjack.py
This one was really fun. I should mention that while it looks like a lot to jump from memory to this, the course staff plugged in a lot of the initial code. We built on top of a template, as with all the projects.

The big lesson in this week was classes, it was an incredible set of lessons.